Move "Create Pool" button to the shared pools list
