+++ src/components/Home/PrivatePools.tsx
@@ -1,19 +1,11 @@
 import React from 'react';
 import styled from 'styled-components';
 import LiquidityPanel, { LiquidityPanelDataSource } from './LiquidityPanel';
-import Button from '../Common/Button';
 import { observer } from 'mobx-react';
-import { Link } from 'react-router-dom';
 import { useStores } from '../../contexts/storesContext';
 
 const Wrapper = styled.div`
     padding-top: 8px;
-`;
-
-const HeaderWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: baseline;
 `;
 
 const Header = styled.div`
@@ -23,36 +15,19 @@ const Header = styled.div`
     font-size: 18px;
     line-height: 19px;
     color: var(--header-text);
-    padding: 0px 0px 24px 0px;
-`;
-
-const CreateLink = styled(Link)`
-    text-decoration: none;
+    padding: 24px 0px 24px 0px;
 `;
 
 const PrivatePools = observer(() => {
     const {
-        root: { poolStore, providerStore },
+        root: { poolStore },
     } = useStores();
 
     const pools = poolStore.getPrivatePools();
 
-    const account = providerStore.providerStatus.account;
-
-    const createButtonHandler = async () => {};
-
     return (
         <Wrapper>
-            <HeaderWrapper>
-                <Header>Private Pools</Header>
-                <CreateLink to={`/pool/new`}>
-                    <Button
-                        buttonText={'Create Pool'}
-                        active={!!account}
-                        onClick={e => createButtonHandler()}
-                    />
-                </CreateLink>
-            </HeaderWrapper>
+            <Header>Private Pools</Header>
             <LiquidityPanel
                 pools={pools}
                 dataSource={LiquidityPanelDataSource.ALL_PUBLIC}
