+++ src/components/Home/SharedPools.tsx
@@ -1,11 +1,19 @@
 import React from 'react';
 import styled from 'styled-components';
 import LiquidityPanel, { LiquidityPanelDataSource } from './LiquidityPanel';
+import Button from '../Common/Button';
 import { observer } from 'mobx-react';
+import { Link } from 'react-router-dom';
 import { useStores } from '../../contexts/storesContext';
 
 const Wrapper = styled.div`
     padding-top: 8px;
+`;
+
+const HeaderWrapper = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: baseline;
 `;
 
 const Header = styled.div`
@@ -15,19 +23,34 @@ const Header = styled.div`
     font-size: 18px;
     line-height: 19px;
     color: var(--header-text);
-    padding: 24px 0px 24px 0px;
+    padding: 0px 0px 24px 0px;
+`;
+
+const CreateLink = styled(Link)`
+    text-decoration: none;
 `;
 
 const SharedPools = observer(() => {
     const {
-        root: { poolStore },
+        root: { poolStore, providerStore },
     } = useStores();
 
     const pools = poolStore.getPublicPools();
 
+    const account = providerStore.providerStatus.account;
+
     return (
         <Wrapper>
-            <Header>Shared Pools</Header>
+            <HeaderWrapper>
+                <Header>Shared Pools</Header>
+                <CreateLink to={`/pool/new`}>
+                    <Button
+                        buttonText={'Create Pool'}
+                        active={!!account}
+                        onClick={e => {}}
+                    />
+                </CreateLink>
+            </HeaderWrapper>
             <LiquidityPanel
                 pools={pools}
                 dataSource={LiquidityPanelDataSource.ALL_PUBLIC}
